QuickAdminPanel automatic commit

+++ database/migrations/2021_03_21_000026_create_permission_role_pivot_table.php
@@ -9,9 +9,9 @@ class CreatePermissionRolePivotTable extends Migration
     public function up()
     {
         Schema::create('permission_role', function (Blueprint $table) {
-            $table->unsignedInteger('role_id');
+            $table->unsignedBigInteger('role_id');
             $table->foreign('role_id', 'role_id_fk_2105623')->references('id')->on('roles')->onDelete('cascade');
-            $table->unsignedInteger('permission_id');
+            $table->unsignedBigInteger('permission_id');
             $table->foreign('permission_id', 'permission_id_fk_2105623')->references('id')->on('permissions')->onDelete('cascade');
         });
     }
